requirements.txt
matplotlib==3.5.3
numpy==1.23.2
pandas==1.4.4
PyYAML==6.0
scikit_learn==1.1.2
streamlit==1.12.2
streamlit_authenticator==0.2.1
streamlit_modal==0.0.5
